requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0

# For Cosmos DB (optional - install if using Cosmos)
azure-cosmos>=4.5.0

# Phase 1.2 & 1.3: Clustering & Analysis
scikit-learn>=1.3.0

# Phase 1.3: Embeddings (choose one or more)
sentence-transformers>=2.2.0  # Local embeddings (recommended for testing)
# openai>=1.0.0               # OpenAI embeddings (requires API key)
# cohere>=4.0.0               # Cohere embeddings (requires API key)

# Phase 1.3: Visualization
umap-learn>=0.5.0            # For UMAP dimensionality reduction
matplotlib>=3.7.0            # For plotting
seaborn>=0.12.0              # For statistical plots
plotly>=5.0.0                # For interactive plots

# Jupyter notebook support
jupyter>=1.0.0
ipywidgets>=8.0.0            # For interactive widgets
